feat(ts-errors): rename
