+++ packages/falcon-dev-tools/falcon-scripts/bin/falcon-scripts.js
@@ -23,11 +23,11 @@ const Logger = require('@deity/falcon-logger');
             .pop()
             .toUpperCase() || undefined;
 
+        const buildDts = require('../src/build-dts');
         const buildEsm = require('../src/build-esm');
-        const buildTsDeclarations = require('../src/build-tsDeclarations');
         const buildCjs = require('../src/build-cjs');
 
-        buildTsDeclarations({ packagePath });
+        buildDts({ packagePath });
         buildEsm({ packagePath, target });
         if (target !== 'NODE') {
           await buildCjs({ packagePath, target });
